chore: refactor customerQuota utility functions to utils
