+++ src/components/CustomerQuota/NoQuota/NoQuotaCard.tsx
@@ -4,7 +4,7 @@ import React, {
   useContext,
   useEffect,
 } from "react";
-import { differenceInSeconds, compareDesc } from "date-fns";
+import { differenceInSeconds } from "date-fns";
 import { View, StyleSheet, ActivityIndicator } from "react-native";
 import { CustomerCard } from "../CustomerCard";
 import { AppText } from "../../Layout/AppText";
@@ -15,11 +15,13 @@ import { Cart } from "../../../hooks/useCart/useCart";
 import { usePastTransaction } from "../../../hooks/usePastTransaction/usePastTransaction";
 import { FontAwesome } from "@expo/vector-icons";
 import {
-  formatQuantityText,
   BIG_NUMBER,
-  sortTransactionsByOrder,
+  groupTransactionsByCategory,
+  sortTransactionsByCategory,
+  getLatestTransactionTime,
+  checkHasAppealProduct,
 } from "../utils";
-import { TransactionsGroup, Transaction } from "../TransactionsGroup";
+import { TransactionsGroup } from "../TransactionsGroup";
 import { ShowFullListToggle } from "../ShowFullListToggle";
 import {
   DistantTransactionTitle,
@@ -28,16 +30,11 @@ import {
   UsageQuotaTitle,
 } from "./TransactionTitle";
 import { AppealButton } from "./AppealButton";
-import { getIdentifierInputDisplay } from "../../../utils/getIdentifierInputDisplay";
-import { Quota, PastTransactionsResult, CampaignPolicy } from "../../../types";
+import { Quota } from "../../../types";
 import { AlertModalContext } from "../../../context/alert";
 import { CampaignConfigContext } from "../../../context/campaignConfig";
 import { AuthContext } from "../../../context/auth";
-import { formatDateTime } from "../../../utils/dateTimeFormatter";
-import {
-  TranslationHook,
-  useTranslate,
-} from "../../../hooks/useTranslate/useTranslate";
+import { useTranslate } from "../../../hooks/useTranslate/useTranslate";
 
 const DURATION_THRESHOLD_SECONDS = 60 * 10; // 10 minutes
 const MAX_TRANSACTIONS_TO_DISPLAY = 5;
@@ -49,14 +46,6 @@ export const styles = StyleSheet.create({
   },
 });
 
-export interface TransactionsByCategoryMap {
-  [category: string]: {
-    transactions: Transaction[];
-    order: number;
-    hasLatestTransaction: boolean;
-  };
-}
-
 interface NoQuotaCard {
   ids: string[];
   cart: Cart;
@@ -65,131 +54,6 @@ interface NoQuotaCard {
   quotaResponse?: Quota;
   allQuotaResponse?: Quota;
 }
-
-/**
- * Given past transactions, group them by categories.
- *
- * @param sortedTransactions Past transaction results sorted by transaction time in desc order
- * @param allProducts Policies
- * @param latestTransactionTime Transaction time of latest transaction
- */
-export const groupTransactionsByCategory = (
-  sortedTransactions: PastTransactionsResult["pastTransactions"] | null,
-  allProducts: CampaignPolicy[],
-  latestTransactionTime: Date | undefined,
-  translationProps: TranslationHook
-): TransactionsByCategoryMap => {
-  const { c13nt, c13ntForUnit, i18nt } = translationProps;
-
-  // Group transactions by category
-  const transactionsByCategoryMap: TransactionsByCategoryMap = {};
-  sortedTransactions?.forEach((item) => {
-    const policy = allProducts?.find(
-      (policy) => policy.category === item.category
-    );
-    const tName = (policy?.name && c13nt(policy?.name)) ?? item.category;
-    const formattedDate = formatDateTime(item.transactionTime);
-
-    if (!transactionsByCategoryMap.hasOwnProperty(tName)) {
-      transactionsByCategoryMap[tName] = {
-        transactions: [],
-        hasLatestTransaction: false,
-        order: policy?.order ?? BIG_NUMBER,
-      };
-    }
-    transactionsByCategoryMap[tName].transactions.push({
-      header: formattedDate,
-      details: getIdentifierInputDisplay(item.identifierInputs ?? []),
-      quantity: formatQuantityText(
-        item.quantity,
-        policy?.quantity.unit
-          ? c13ntForUnit(policy?.quantity.unit)
-          : {
-              type: "POSTFIX",
-              label: ` ${i18nt("checkoutSuccessScreen", "quantity")}`,
-            }
-      ),
-      isAppeal: policy?.categoryType === "APPEAL",
-      order: -1,
-    });
-    transactionsByCategoryMap[tName].hasLatestTransaction =
-      transactionsByCategoryMap[tName].hasLatestTransaction ||
-      differenceInSeconds(latestTransactionTime || 0, item.transactionTime) ===
-        0;
-  });
-
-  return transactionsByCategoryMap;
-};
-
-/**
- * Transforms map of transactions into an array
- * Array is sorted by:
- *  1. Categories with latest transactions
- *  2. All other categories
- * Each group is sorted by category name in asc order
- * Also adds an order attribute to each transaction for the "Show more" feature
- *
- * @param transactionsByCategoryMap Transactions by category
- */
-export const sortTransactions = (
-  transactionsByCategoryMap: TransactionsByCategoryMap
-): TransactionsGroup[] => {
-  const latestTransactionsByCategory: TransactionsGroup[] = [];
-  const otherTransactionsByCategory: TransactionsGroup[] = [];
-  Object.entries(transactionsByCategoryMap).forEach(([key, value]) => {
-    const { transactions, hasLatestTransaction, order } = value;
-    if (hasLatestTransaction) {
-      latestTransactionsByCategory.push({
-        header: key,
-        transactions,
-        order,
-      });
-    } else {
-      otherTransactionsByCategory.push({ header: key, transactions, order });
-    }
-  });
-
-  latestTransactionsByCategory.sort(sortTransactionsByOrder);
-  otherTransactionsByCategory.sort(sortTransactionsByOrder);
-
-  let transactionCounter = 0;
-  return latestTransactionsByCategory
-    .concat(otherTransactionsByCategory)
-    .map((transactionsByCategory) => {
-      const orderedTransactions = transactionsByCategory.transactions.map(
-        (transaction) => {
-          const result = { ...transaction, order: transactionCounter };
-          transactionCounter += 1;
-          return result;
-        }
-      );
-      return { ...transactionsByCategory, transactions: orderedTransactions };
-    });
-};
-
-export const getLatestTransactionTime = (cart: Cart): Date | undefined =>
-  cart.sort((item1, item2) =>
-    compareDesc(item1.lastTransactionTime ?? 0, item2.lastTransactionTime ?? 0)
-  )[0]?.lastTransactionTime;
-
-export const checkHasAppealProduct = (
-  allProducts: CampaignPolicy[] | null,
-  allQuotaResponse: Quota | undefined
-): boolean => {
-  const appealProductsCategories = allProducts
-    ?.filter((product) => product.categoryType === "APPEAL")
-    .map((product) => product.category);
-
-  const hasAppealProduct =
-    (appealProductsCategories &&
-      allQuotaResponse?.remainingQuota.some(
-        (quota) =>
-          appealProductsCategories.includes(quota.category) &&
-          quota.quantity !== 0
-      )) ??
-    false;
-  return hasAppealProduct;
-};
 
 /**
  * Shows when the user cannot purchase anything
@@ -242,7 +106,7 @@ export const NoQuotaCard: FunctionComponent<NoQuotaCard> = ({
     translationProps
   );
 
-  const transactionsByCategoryList = sortTransactions(
+  const transactionsByCategoryList = sortTransactionsByCategory(
     transactionsByCategoryMap
   );
